more warning squashing

See merge request graphviz/graphviz!2052

+++ lib/sfio/sffilbuf.c
@@ -9,6 +9,7 @@
  *************************************************************************/
 
 #include	<sfio/sfhdr.h>
+#include	<stddef.h>
 
 /*	Fill the buffer of a stream with data.
 **	If n < 0, sffilbuf() attempts to fill the buffer if it's empty.
@@ -55,7 +56,7 @@ int _sffilbuf(Sfio_t * f, int n)
 	    /* try shifting left to make room for new data */
 	    if (!(f->bits & SF_MMAP) && f->next > f->data &&
 		n > (f->size - (f->endb - f->data))) {
-		memcpy(f->data, f->next, r);
+		memcpy(f->data, f->next, (size_t)r);
 		f->next = f->data;
 		f->endb = f->data + r;
 	    }
@@ -69,7 +70,7 @@ int _sffilbuf(Sfio_t * f, int n)
 	    if (n > 0) {
 		if (r > n && f->extent < 0 && (f->flags & SF_SHARE))
 		    r = n;	/* read only as much as requested */
-		else if (justseek && n <= f->iosz && f->iosz <= f->size)
+		else if (justseek && (size_t)n <= f->iosz && f->iosz <= f->size)
 		    r = f->iosz;	/* limit buffer filling */
 	    }
 	}
